+++ src/OpenTelemetry.Exporter.InfluxDB/InfluxDBExporterExtensions.cs
@@ -56,14 +56,11 @@ public static class InfluxDBExporterExtensions
         return builder;
     }
 
-    private static IMetricsWriter CreateMetricsWriter(MetricsSchema metricsSchema)
+    private static IMetricsWriter CreateMetricsWriter(MetricsSchema metricsSchema) => metricsSchema switch
     {
-        return metricsSchema switch
-        {
-            MetricsSchema.TelegrafPrometheusV2 => new TelegrafPrometheusWriterV2(),
-            MetricsSchema.TelegrafPrometheusV1 => new TelegrafPrometheusWriterV1(),
-            MetricsSchema.None => new TelegrafPrometheusWriterV1(),
-            _ => new TelegrafPrometheusWriterV1(),
-        };
-    }
+        MetricsSchema.TelegrafPrometheusV2 => new TelegrafPrometheusWriterV2(),
+        MetricsSchema.TelegrafPrometheusV1 => new TelegrafPrometheusWriterV1(),
+        MetricsSchema.None => new TelegrafPrometheusWriterV1(),
+        _ => new TelegrafPrometheusWriterV1(),
+    };
 }