+++ src/OpenTelemetry.Exporter.Instana/Implementation/Processors/ErrorActivityProcessor.cs
@@ -5,9 +5,9 @@ using System.Diagnostics;
 
 namespace OpenTelemetry.Exporter.Instana.Implementation.Processors;
 
-internal class ErrorActivityProcessor : ActivityProcessorBase, IActivityProcessor
+internal sealed class ErrorActivityProcessor : ActivityProcessorBase, IActivityProcessor
 {
-    public override async Task ProcessAsync(Activity activity, InstanaSpan instanaSpan)
+    public override void Process(Activity activity, InstanaSpan instanaSpan)
     {
         this.PreProcess(activity, instanaSpan);
 
@@ -23,18 +23,14 @@ internal class ErrorActivityProcessor : ActivityProcessorBase, IActivityProcesso
                 }
             }
         }
-        else if (instanaSpan.TransformInfo != null && instanaSpan.TransformInfo.HasExceptionEvent)
-        {
-            instanaSpan.Ec = 1;
-        }
         else
         {
-            instanaSpan.Ec = 0;
+            instanaSpan.Ec = instanaSpan.TransformInfo != null && instanaSpan.TransformInfo.HasExceptionEvent ? 1 : 0;
         }
 
         if (activity != null && instanaSpan != null)
         {
-            await base.ProcessAsync(activity, instanaSpan).ConfigureAwait(false);
+            base.Process(activity, instanaSpan);
         }
     }
 }