+++ test/OpenTelemetry.Exporter.Instana.Tests/Processors/DefaultActivityProcessorTests.cs
@@ -10,28 +10,29 @@ namespace OpenTelemetry.Exporter.Instana.Tests.Processors;
 
 public class DefaultActivityProcessorTests
 {
-    private readonly DefaultActivityProcessor defaultActivityProcessor = new();
-
     [Fact]
-    public async Task ProcessAsync()
+    public void Process_PopulatesInstanaSpan()
     {
+        // Arrange
         var activity = new Activity("testOperationName");
         activity.Start();
-        await Task.Delay(200);
+
+        Thread.Sleep(200); // Simulate some work being done
+
         activity.Stop();
         var instanaSpan = new InstanaSpan();
-        await this.defaultActivityProcessor.ProcessAsync(activity, instanaSpan);
 
+        // Act
+        var processor = new DefaultActivityProcessor();
+        processor.Process(activity, instanaSpan);
+
+        // Assert
         Assert.False(string.IsNullOrEmpty(instanaSpan.S));
         Assert.False(string.IsNullOrEmpty(instanaSpan.Lt));
         Assert.True(instanaSpan.D > 0);
         Assert.True(instanaSpan.Ts > 0);
         Assert.NotNull(instanaSpan.Data);
         Assert.NotNull(instanaSpan.Data.data);
-        Assert.Contains(instanaSpan.Data.data, filter: x =>
-        {
-            return x.Key == "kind"
-                   && x.Value.Equals("internal");
-        });
+        Assert.Contains(instanaSpan.Data.data, filter: x => x.Key == "kind" && x.Value.Equals("internal"));
     }
 }